requirements.txt
# Core Dependencies
pandas>=1.3.0
requests>=2.26.0

# Google API Dependencies
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

